dev-requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
fsspec==2024.10.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.1.3
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.3.0
sniffio==1.3.1
tokenizers==0.20.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.1
wheel==0.44.0
